requirements.txt
keyboard==0.13.5
customtkinter==5.2.1
opencv-python==4.9.0.80
pillow==10.1.0
pyautogui==0.9.54
pynput==1.7.6
mediapipe==0.10.8
CTkToolTip==0.8
# pyinstaller==6.3.0